+++ src/background/state/application/events.rs
@@ -9,7 +9,7 @@ impl FullState {
     pub(super) fn emit_settings_by_app(&self) -> Result<()> {
         get_app_handle().emit(
             SeelenEvent::StateSettingsByAppChanged,
-            &self.settings_by_app,
+            self.settings_by_app.as_slice(),
         )?;
         Ok(())
     }